Allow customizing ElasticSearch index name in UTC or local time.
Also allow customizing ElasticSearch index format.

These changes allow for better integration with existing log analysis tools, such as ElasticSearch's Kibana.

Note: The default behavior remained the same as in 2.3 and 2.4. Depending on feedback, we might want to change the default for 2.5. Examples for better interoperability with Kibana are suggested in the documentation.

Addresses BIT-1336