Fix failing elasticsearch tests due to netcat
Added a shell script wrapper to choose the correct cmd-line arguments
to the nc command.  This is needed because some systems (such as debian 7
and 8) use an implementation of netcat that requires the listening
port number to be specified with a "-p" cmd-line argument, which is not
allowed by the implementation of netcat used on other systems.